feat: 사용자이름, 한줄소개 수정 UI 및 navigate
