+++ screens/index.js
@@ -1,3 +1,5 @@
 export { Search } from './Search';
 export { TimeLine } from './TimeLine';
 export { EditProfile } from './Profile/EditProfile';
+export { EditUsername } from './Profile/EditUsername';
+export { EditIntroduction } from './Profile/EditIntroduction';
